Separated types
This allows for an easier implementation of type-family based queries
in the near future. Such a model allows for parametrizing queries over
the type of AST we use, ruling out useless queries like searches for
OOP-style classes in Haskell code or for macros in python. This does not
mean, however, that all representable queries will yield a result, for
instance, ranges like 3-1 aren't caught that way, but they are a single
corner case that is easy to rule out.