Added recursive lookup of class declarations
When querying for a type or function, class methods and associated
types and data families are also looked up. This throws up the question
whether we want to keep a special query type to search for families,
or whether we should merge into type queries, as those two can't be
distinguished by the user when he looks at a bit of code.